Fixed role deletion bug
The many-to-many configuration in the book is somewhat obsolete.
Many-to-many objects should have a collections of each on both sides.
Employee should have getRoles() and Roles should have getEmployees()

Improved UI, in add roles to employee.

Employee will only see the list of roles that does not belong to
its roleset.